map functions reflect current patterning
