+++ campsite-detail.html
@@ -14,7 +14,6 @@
     <header id="header-container">
     </header>
     <main>
-        <img src="" id="googlepic" alt="">
         <section id="main">
         </section>
         <section id="forecast-display">
